refactoring of package hierarchy and cleanup
