+++ adbms/src/main/scala/de/up/hpi/informationsystems/adbms/Dactor.scala
@@ -4,6 +4,7 @@ import akka.actor.{Actor, ActorLogging, ActorRef, ActorRefFactory, ActorSelectio
 import akka.util.Timeout
 import de.up.hpi.informationsystems.adbms.definition._
 import de.up.hpi.informationsystems.adbms.protocols.RequestResponseProtocol
+import de.up.hpi.informationsystems.adbms.relation.{FutureRelation, MutableRelation, Relation, RowRelation}
 
 import scala.concurrent.Future
 
@@ -107,7 +108,7 @@ object Dactor {
 
   /**
     * Constructs a mapping of relation definitions and corresponding relational stores using
-    * [[de.up.hpi.informationsystems.adbms.definition.RowRelation]] as base relation.
+    * [[de.up.hpi.informationsystems.adbms.relation.RowRelation]] as base relation.
     *
     * @param relDefs sequence of relation definitions
     * @return mapping of relation definition and corresponding relational row store
