+++ adbms/src/main/scala/de/up/hpi/informationsystems/adbms/protocols/RequestResponseProtocol.scala
@@ -1,6 +1,6 @@
 package de.up.hpi.informationsystems.adbms.protocols
 
-import de.up.hpi.informationsystems.adbms.definition.Record
+import de.up.hpi.informationsystems.adbms.record.Record
 
 object RequestResponseProtocol {
