+++ adbms/src/main/scala/de/up/hpi/informationsystems/adbms/record/ColumnCellMapping.scala
@@ -1,10 +1,12 @@
-package de.up.hpi.informationsystems.adbms.definition
+package de.up.hpi.informationsystems.adbms.record
+
+import de.up.hpi.informationsystems.adbms.definition.{ColumnDef, UntypedColumnDef}
 
 /**
   * Provides implicits for creating a mapping between column definition and cell value
   *
   * @example {{{
-  * import de.up.hpi.informationsystems.adbms.definition.ColumnCellMapping._
+  * import de.up.hpi.informationsystems.adbms.record.ColumnCellMapping._
   * val x: ColumnCellMapping = ColumnDef[Int]("id") ~> 12
   * }}}
   */
@@ -13,22 +15,23 @@ object ColumnCellMapping {
 
     /**
       * Syntax-sugar for creating a mapping of column definition and cell value for the use in e.g.
-      * [[de.up.hpi.informationsystems.adbms.definition.Record.RecordBuilder]]. Returns a
-      * [[de.up.hpi.informationsystems.adbms.definition.ColumnCellMapping]].
+      * [[de.up.hpi.informationsystems.adbms.record.Record.RecordBuilder]]. Returns a
+      * [[de.up.hpi.informationsystems.adbms.record.ColumnCellMapping]].
+      *
       * @param value cell value
-      * @return a new [[de.up.hpi.informationsystems.adbms.definition.ColumnCellMapping]] containing
+      * @return a new [[de.up.hpi.informationsystems.adbms.record.ColumnCellMapping]] containing
       *         the column definition and cell value mapping
       */
     def ~>(value: T): ColumnCellMapping = new ColumnCellMapping(Map(in.untyped -> value))
   }
 }
 
-final class ColumnCellMapping private[definition](private val columnCellMapping: Map[UntypedColumnDef, Any]){
+final class ColumnCellMapping private[adbms](private val columnCellMapping: Map[UntypedColumnDef, Any]){
   def +(other: ColumnCellMapping): ColumnCellMapping = and(other)
   def &(other: ColumnCellMapping): ColumnCellMapping = and(other)
   def and(other: ColumnCellMapping): ColumnCellMapping =
     new ColumnCellMapping(this.columnCellMapping ++ other.toMap)
 
-  private[definition] def toMap: Map[UntypedColumnDef, Any] = columnCellMapping
+  private[adbms] def toMap: Map[UntypedColumnDef, Any] = columnCellMapping
 }
