+++ adbms/src/test/scala/de/up/hpi/informationsystems/adbms/csv/CSVParserTest.scala
@@ -3,9 +3,10 @@ package de.up.hpi.informationsystems.adbms.csv
 import java.io._
 import java.net.URI
 
-import de.up.hpi.informationsystems.adbms.definition.ColumnCellMapping._
 import de.up.hpi.informationsystems.adbms.definition.ColumnTypeDefaults._
-import de.up.hpi.informationsystems.adbms.definition.{ColumnDef, Relation, RelationDef, UntypedColumnDef}
+import de.up.hpi.informationsystems.adbms.definition.{ColumnDef, RelationDef, UntypedColumnDef}
+import de.up.hpi.informationsystems.adbms.record.ColumnCellMapping._
+import de.up.hpi.informationsystems.adbms.relation.Relation
 import org.scalatest.{Matchers, WordSpec}
