+++ sampleapp/src/test/scala/de/up/hpi/informationsystems/sampleapp/test/dactors/SystemTest.scala
@@ -5,10 +5,11 @@ import java.time.{ZoneOffset, ZonedDateTime}
 import akka.actor.ActorSystem
 import akka.testkit.{TestKit, TestProbe}
 import de.up.hpi.informationsystems.adbms.Dactor
-import de.up.hpi.informationsystems.adbms.definition.ColumnCellMapping._
+import de.up.hpi.informationsystems.adbms.definition.ColumnDef
 import de.up.hpi.informationsystems.adbms.definition.ColumnTypeDefaults._
-import de.up.hpi.informationsystems.adbms.definition.{ColumnDef, Record}
 import de.up.hpi.informationsystems.adbms.protocols.DefaultMessagingProtocol.InsertIntoRelation
+import de.up.hpi.informationsystems.adbms.record.ColumnCellMapping._
+import de.up.hpi.informationsystems.adbms.record.Record
 import de.up.hpi.informationsystems.sampleapp.dactors.Cart.{CartInfo, CartPurchases}
 import de.up.hpi.informationsystems.sampleapp.dactors.Customer.{CustomerInfo, StoreVisits}
 import de.up.hpi.informationsystems.sampleapp.dactors.GroupManager.Discounts
